+++ tests/test_job_scheduler.ts
@@ -340,6 +340,14 @@ describe('Job Scheduler', function () {
     });
   });
 
+  describe('when job scheduler does not exist', function () {
+    it('should return undefined', async function () {
+      const scheduler = await queue.getJobScheduler('test');
+
+      expect(scheduler).to.be.undefined;
+    });
+  });
+
   it('should create job schedulers with different cron patterns', async function () {
     const date = new Date('2017-02-07T15:24:00.000Z');
     this.clock.setSystemTime(date);
@@ -463,10 +471,7 @@ describe('Job Scheduler', function () {
     expect(scheduler).to.deep.equal({
       key: 'test',
       name: 'test',
-      endDate: null,
-      tz: null,
       pattern: '*/2 * * * * *',
-      every: null,
       next: 1486481042000,
       template: {
         data: {
@@ -805,11 +810,8 @@ describe('Job Scheduler', function () {
         expect(scheduler).to.deep.equal({
           key: 'rrule',
           name: 'rrule',
-          endDate: null,
           next: 1486481042000,
-          tz: null,
           pattern: 'RRULE:FREQ=SECONDLY;INTERVAL=2;WKST=MO',
-          every: null,
         });
 
         this.clock.tick(nextTick);
@@ -1715,10 +1717,8 @@ describe('Job Scheduler', function () {
       expect(jobSchedulers[0]).to.deep.equal({
         key: 'test',
         name: 'a',
-        endDate: null,
         tz: 'Asia/Calcutta',
         pattern: '0 * 1 * *',
-        every: null,
         next: 1488310200000,
         template: {
           data: {
@@ -2268,10 +2268,8 @@ describe('Job Scheduler', function () {
     const scheduler = await queue.getJobScheduler('repeat');
 
     expect(scheduler).to.deep.equal({
-      endDate: null,
-      every: null,
-      tz: null,
       key: 'repeat',
+      limit: 5,
       name: 'repeat',
       pattern: '*/1 * * * * *',
       next: 1486439641000,